+++ changelog.md
@@ -62,3 +62,5 @@
 - Now it can be considered usable and we plan to maintain and improve it in nearest future with more options, languages, previews, shortcuts!
 - To do: fix bugs, make more options, refactor the code
 - Added dynamic color picker for the preview
+- Auto extension detector
+- Default extension